requirements.txt
Flask==3.0.2
opencv-python-headless==4.9.0.80
scikit-image==0.22.0
yt-dlp==2025.2.19
numpy==1.26.4
fpdf2==2.7.8
Pillow==10.2.0
